requirements.txt
Django==3.1.3
djangorestframework==3.12.2
drf-yasg==1.20.0
django-cors-headers==3.5.0
drf-yasg==1.20.0
celery[redis]==5.1.2
django-celery-beat==2.2.1
psycopg2-binary>=2.9.1,<2.9.2
requests==2.7.0
amqp==5.0.6
boto3==1.21.29
django-celery-results==2.2.0